Move Command, CommandInterpreter and Command parser into single command module
