+++ contextshell/command.py
@@ -1,5 +1,51 @@
-from contextshell.Command import Command
 from typing import List, Optional
+
+from contextshell.NodePath import NodePath
+from contextshell.ActionExecutor import ActionExecutor, parse_argument_tree
+
+
+class Command:
+    """Represents single command line typed in the shell"""
+
+    def __init__(self, command_name):
+        self.target = None
+        self.name = command_name
+        self.arguments = []
+
+    def __str__(self):
+        representation = " ".join(map(Command._to_string, [self.name] + self.arguments))
+        if self.target is None:
+            return representation
+        representation = "{}: {}".format(Command._to_string(self.target), representation)
+        return representation
+
+    @staticmethod
+    def _to_string(param):
+        if isinstance(param, Command):
+            return "{{{}}}".format(str(param))
+        return str(param)
+
+
+class CommandInterpreter:
+    def __init__(self, tree: ActionExecutor) -> None:
+        self.tree = tree
+
+    def execute(self, command: Command):
+        if command is None:
+            raise ValueError("No command to execute provided")
+        target_path = self._evaluate(command.target)
+        if target_path is None:
+            raise RuntimeError("No action target specified")
+        target_path = NodePath.cast(target_path)
+        action_path = NodePath.cast(self._evaluate(command.name))
+        arguments = list(map(self._evaluate, command.arguments))
+        packed_arguments = parse_argument_tree(arguments)
+        return self.tree.execute(target_path, action_path, packed_arguments)
+
+    def _evaluate(self, part):
+        if isinstance(part, Command):
+            return self.execute(part)
+        return part
 
 
 def convert_token_type(token):
